Arreglos en arrancar y detener

+++ detener.sh
@@ -4,14 +4,20 @@ GRALOG="./GraLog.sh"
 
 PID=$(pgrep "$1")
 
-if [[ "$PID" != "" ]]
+if [ "$#" -eq 1 ]
 then
-        echo Se detuvo el proceso $1
-        $GRALOG "detener" "Se detuvo el proceso $1" "INFO"
-	kill -9 $PID
-       # exit 0
-else
-	echo "No existe el proceso"
-	$GRALOG "detener" "No existe el proceso" "ERR"
-fi
 
+	if [[ "$PID" != "" ]]
+	then
+	        echo Se detuvo el proceso $1
+	        $GRALOG "detener" "Se detuvo el proceso $1" "INFO"
+			kill -9 $PID
+
+	else
+		echo "No existe el proceso"
+		$GRALOG "detener" "No existe el proceso" "ERR"
+	fi
+else
+	echo "Se debe indicar un proceso univoco a detener"
+	$GRALOG "detener" "Se debe indicar un proceso univoco a detener" "ERR"
+fi
